+++ src/main/java/me/kall/duplicationless/ext/DataRebuilder.java
@@ -1,0 +1,6 @@
+package me.kall.duplicationless.ext;
+
+public interface DataRebuilder {
+    boolean duplicationless$rebuilt();
+    void duplicationless$setRebuilt();
+}