+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # 1.0.5
 - Optimize ChunkData#rebuild for much faster execution.
+- introduce rebuildChunk for chunk-specific data rebuilding.